requirements.txt
sqlalchemy
zope.sqlalchemy
celery
